+++ integration/rabbitmq/e2e/configuration.e2e-spec.ts
@@ -2,8 +2,9 @@ import { RabbitMQConfig, RabbitMQModule } from '@golevelup/nestjs-rabbitmq';
 import { Test, TestingModule } from '@nestjs/testing';
 import * as amqplib from 'amqplib';
 
-const rabbitHost = process.env.NODE_ENV === 'ci' ? 'rabbit' : 'localhost';
-const uri = `amqp://rabbitmq:rabbitmq@${rabbitHost}:5672`;
+const rabbitHost = process.env.NODE_ENV === 'ci' ? process.env.RABBITMQ_HOST : 'localhost';
+const rabbitPort = process.env.NODE_ENV === 'ci' ? process.env.RABBITMQ_PORT : '5672';
+const uri = `amqp://rabbitmq:rabbitmq@${rabbitHost}:${rabbitPort}`;
 const amqplibUri = `${uri}?heartbeat=5`;
 
 class RabbitConfig {
@@ -19,7 +20,10 @@ class RabbitConfig {
 describe('Module Configuration', () => {
   let app: TestingModule;
 
-  afterEach(() => jest.clearAllMocks());
+  afterEach(async () => {
+    jest.clearAllMocks();
+    await app.close();
+  });
 
   describe('forRoot', () => {
     it('should configure RabbitMQ', async () => {