+++ packages/stripe/src/tests/stripe.module.spec.ts
@@ -1,10 +1,13 @@
-import { INestApplication, Injectable } from '@nestjs/common';
+import { INestApplication, Injectable, SetMetadata } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import Stripe from 'stripe';
 import { InjectStripeClient } from '../stripe.decorators';
+import { StripeWebhookController } from '../stripe.webhook.controller';
 import { StripeModule } from './../stripe.module';
 
 const testReceiveStripeFn = jest.fn();
+
+const TestDecorator = () => SetMetadata('TEST:METADATA', 'metadata');
 
 @Injectable()
 class TestService {
@@ -35,5 +38,21 @@ describe('Stripe Module', () => {
 
     const client = testReceiveStripeFn.mock.calls[0][0];
     expect(client).toBeInstanceOf(Stripe);
+  });
+  it('should apply the decorator to the controller', async () => {
+    await Test.createTestingModule({
+      imports: [
+        StripeModule.forRoot(StripeModule, {
+          apiKey: '123',
+          webhookConfig: {
+            stripeWebhookSecret: 'super-secret',
+            decorators: [TestDecorator()],
+          },
+        }),
+      ],
+    }).compile();
+    expect(Reflect.getMetadata('TEST:METADATA', StripeWebhookController)).toBe(
+      'metadata'
+    );
   });
 });