+++ packages/hasura/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@1.3.0...@golevelup/nestjs-hasura@1.3.1) (2021-10-06)
+
+**Note:** Version bump only for package @golevelup/nestjs-hasura
+
+# [1.3.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@1.2.1...@golevelup/nestjs-hasura@1.3.0) (2021-07-06)
+
+### Features
+
+- **hasura:** metadata v3 support ([#284](https://github.com/golevelup/nestjs/issues/284)) ([bcb6fc6](https://github.com/golevelup/nestjs/commit/bcb6fc6)), closes [#250](https://github.com/golevelup/nestjs/issues/250)
+
+## [1.2.1](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@1.2.0...@golevelup/nestjs-hasura@1.2.1) (2021-05-17)
+
+**Note:** Version bump only for package @golevelup/nestjs-hasura
+
+# [1.2.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@1.1.0...@golevelup/nestjs-hasura@1.2.0) (2021-03-02)
+
+### Features
+
+- **hasura:** allow for custom decorators on controller ([32802e6](https://github.com/golevelup/nestjs/commit/32802e6))
+
+# [1.1.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@1.0.0...@golevelup/nestjs-hasura@1.1.0) (2021-01-16)
+
+### Features
+
+- **hasura:** metadata output property ordering ([9e15b1b](https://github.com/golevelup/nestjs/commit/9e15b1b)), closes [#219](https://github.com/golevelup/nestjs/issues/219)
+
+# [1.0.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.6.0...@golevelup/nestjs-hasura@1.0.0) (2020-12-21)
+
+### Features
+
+- **hasura:** managed event and cron triggers ([92b74fa](https://github.com/golevelup/nestjs/commit/92b74fa))
+
+### BREAKING CHANGES
+
+- **hasura:** new format for Hasura module configuration
+
+# [0.6.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.5.0...@golevelup/nestjs-hasura@0.6.0) (2020-09-02)
+
+### Features
+
+- **hasura:** scheduled event support ([d677854](https://github.com/golevelup/nestjs/commit/d677854)), closes [#175](https://github.com/golevelup/nestjs/issues/175)
+
+# [0.5.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.4.0...@golevelup/nestjs-hasura@0.5.0) (2020-06-10)
+
+### Features
+
+- **hasura:** add HasuraAction type ([0b51c46](https://github.com/golevelup/nestjs/commit/0b51c46))
+
+# [0.4.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.3.3...@golevelup/nestjs-hasura@0.4.0) (2020-05-24)
+
+### Bug Fixes
+
+- **hasura:** mismatch on variables ([592b14e](https://github.com/golevelup/nestjs/commit/592b14e))
+
+### Features
+
+- **hasura:** enable event routing based on trigger name ([877462c](https://github.com/golevelup/nestjs/commit/877462c)), closes [#152](https://github.com/golevelup/nestjs/issues/152)
+
+## [0.3.3](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.3.2...@golevelup/nestjs-hasura@0.3.3) (2020-05-13)
+
+### Bug Fixes
+
+- **stripe, hasura:** forRootAsync dynamic controllers ([c4d10bf](https://github.com/golevelup/nestjs/commit/c4d10bf)), closes [#148](https://github.com/golevelup/nestjs/issues/148)
+
+## [0.3.1-rc.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.3.0...@golevelup/nestjs-hasura@0.3.1-rc.0) (2020-04-08)
+
+### Features
+
+- **stripe:** new stripe package ([ffbcc86](https://github.com/golevelup/nestjs/commit/ffbcc86))
+
 # [0.3.0](https://github.com/golevelup/nestjs/compare/@golevelup/nestjs-hasura@0.2.1...@golevelup/nestjs-hasura@0.3.0) (2020-04-05)
 
 ### Features